linked_list-test: improvements to debuggability of failures
I am working on reproducing and debugging an issue where after running
linked_list-test for many hours I eventually have it fail with 10 or so missing
rows (likely to be one missing write operation). The failure only results in
the count being too low, rather than any missing link in the linked list chain,
which indicates it's probably the last write operation in one of the tablets
going missing.

This patch tries to make that easier to debug in two ways:

* Log the last written chain entries upon completion of insertion. This will
  give us some keys to search for within the stored data.

* Add an option to bootstrap to _not_ delete the replayed logs, even if replay
  was successful. This should help diagnose if in fact the original commit was
  logged but somehow it got lost on replay.

In order to do the latter, I added some new plumbing to external_mini_cluster
to pass arbitrary flags into the TS process.

Reviewed-on: http://gerrit.ent.cloudera.com:8080/2603
Tested-by: jenkins

+++ src/tablet/tablet_bootstrap.cc
@@ -3,6 +3,7 @@
 #include "tablet/tablet_bootstrap.h"
 
 #include <boost/foreach.hpp>
+#include <gflags/gflags.h>
 #include <string>
 #include <utility>
 #include <vector>
@@ -60,6 +61,9 @@ using kudu::tserver::ChangeConfigRequestPB;
 using kudu::tserver::WriteRequestPB;
 using std::tr1::shared_ptr;
 using strings::Substitute;
+
+DEFINE_bool(skip_remove_old_recovery_dir, false,
+            "Skip removing WAL recovery dir after startup. (useful for debugging)");
 
 namespace kudu {
 namespace tablet {
@@ -422,6 +426,11 @@ Status TabletBootstrap::RemoveRecoveryDir() {
                         Substitute("Could not rename old recovery dir from: $0 to: $1",
                                    recovery_path, tmp_path));
   LOG(INFO) << "Renamed old recovery dir from: "  << recovery_path << " to: " << tmp_path;
+
+  if (FLAGS_skip_remove_old_recovery_dir) {
+    LOG(INFO) << "--skip_remove_old_recovery_dir enabled. NOT removing " << tmp_path;
+    return Status::OK();
+  }
   RETURN_NOT_OK_PREPEND(fs_manager->env()->DeleteRecursively(tmp_path),
                         "Could not remove renamed recovery dir" +  tmp_path);
   LOG(INFO) << "Removed renamed recovery dir: " << tmp_path;
